requirements.txt
asgiref==3.8.1
Django==5.0.6
django-rest==0.8.7
psycopg2==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0

djangorestframework~=3.15.2